Renamed and moved things around further and fixed tests.
