+++ EventStream/transformer/lightning_modules/fine_tuning.py
@@ -27,17 +27,17 @@ from torchmetrics.classification import (
 )
 from transformers import get_polynomial_decay_schedule_with_warmup
 
+from ...data.pytorch_dataset import PytorchDataset
+from ...utils import hydra_dataclass, task_wrapper
+from ..config import OptimizationConfig, StructuredTransformerConfig
 from ..data.config import (
     PytorchDatasetConfig,
     SeqPaddingSide,
     SubsequenceSamplingStrategy,
 )
-from ..data.pytorch_dataset import PytorchDataset
-from ..utils import hydra_dataclass, task_wrapper
-from .config import OptimizationConfig, StructuredTransformerConfig
-from .model import ESTForStreamClassification
-from .model_output import StreamClassificationModelOutput
-from .utils import str_summary
+from ..fine_tuning_model import ESTForStreamClassification
+from ..model_output import StreamClassificationModelOutput
+from ..utils import str_summary
 
 
 class ESTForStreamClassificationLM(L.LightningModule):
