+++ scripts/pretrain.py
@@ -14,7 +14,7 @@ import hydra
 import torch
 from omegaconf import OmegaConf
 
-from EventStream.transformer.generative_sequence_modelling_lightning import (
+from EventStream.transformer.lightning_modules.generative_modelling import (
     PretrainConfig,
     train,
 )
