+++ scripts/zeroshot.py
@@ -10,7 +10,10 @@ except ImportError:
 import hydra
 import torch
 
-from EventStream.transformer.zero_shot_utils import FinetuneConfig, zero_shot_evaluation
+from EventStream.transformer.lightning_modules.zero_shot_evaluator import (
+    FinetuneConfig,
+    zero_shot_evaluation,
+)
 
 torch.set_float32_matmul_precision("high")
